Simplified install by running software updates last
Sometimes softwareupdate will require a reboot. I've observed that brew
will sometimes fail to install things if the machine is in that state.
So might as well brew install the things before the softwareudpater
runs. The notification to reboot the machine will be at the end of the
output that way. Also, if the softwareupdates fail, at least the machine
is in working order.